+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 * :feature:`7146` The exported CSV for PnL Report now contains an Asset column with symbols.
 * :feature:`6254` Users can now stop the execution of long-running queries.
 * :feature:`7092` Users of metamask swaps will now see them properly decoded in the history view and have them taken into account during accounting.
+* :feature:`-`  Bridging transactions across all EVM chains done via socket will now be decoded properly.
 * :feature:`7051` Users will be able to add a custom price directly from the asset overview page.
 * :feature:`4841` The rotki data directory is now more organized.
 * :bug:`7147` rotki should no longer query price multiple times for the same asset across different chains.